HZSP/requirements.txt
flask==3.0.2
openai==1.87.0
python-dotenv==1.0.1 